이영경 : detail. logout관련 파일들 수정 20200810 12:25 commit
