--- src/main/java/com/drumtong/business/dao/BusinessDAO.java
@@ -1,7 +1,0 @@
-package com.drumtong.business.dao;
-
-public interface BusinessDAO {
-	String test();
-	//test
-	
-}
