--- src/main/java/com/drumtong/business/vo/BusinessVO.java
@@ -1,5 +1,0 @@
-package com.drumtong.business.vo;
-
-public class BusinessVO {
-
-}
