+++ src/main/java/com/drumtong/customer/service/membership/CustomerMembershipService.java
@@ -87,7 +87,7 @@ public class CustomerMembershipService {
 	// 로그아웃 객체
 	public ModelAndView logout(HttpServletRequest req, HttpServletResponse resp) {
 		String Referer = req.getHeader("referer");
-		String AddressToMove = Referer.split("drumtong")[1];
+		String AddressToMove = Referer != null ? Referer.split("drumtong")[1] : "/customer/";
 		ModelAndView mav = new ModelAndView("redirect:" + AddressToMove);
 		HttpSession Session = req.getSession();
 		SLoginLogVO sLoginLogVO = new SLoginLogVO();
