+++ src/main/webapp/resources/customer/css/customerStyleDohyun.css
@@ -81,7 +81,7 @@ section {
 body {
 	/* border: 1px solid black; */
 	margin: auto;
-	height: 1500px;
+	height: auto;
 
 	/* 	최대크기 */
 	max-width: 1900px;
@@ -95,7 +95,7 @@ body {
 /* ====================섹션 CSS================================ */
 body section {
 	width: 1900px;
-	height: 1400px;
+	height: auto;
 	/* padding: 10px; */
 	/* margin-top: 70px; */
 }
